start deprecating current manager module for it's own module
